Change: reminderState to two separate stateflows
Previously, the reminderState stateflow was being used in both the
'Completed' and 'Incomplete' screens. This was causing a sort of race
condition where a Flow emitted by the Room database would cause both the
getIncompleteReminders and getCompleteReminders methods to update the
same stateflow simultaneously. This resulted in a bug where different
lists would be displayed, without navigating to a different screen,
after clicking the radio button to mark a reminder as complete