Fix: non-responsive due datetime in CreateScreen
Previously, the dueDate and dueTime string displayed in the
CreateReminderScreen would not respond to changes in the state. This was
because the formattedDate and formattedTime variables would only be
assigned the state values once. This would not cause any recomposition
of the Composable. This was changed by assigning the values directly
instead of using the delegate remember function